Moved AzureDatalake Gen1 out of spec.py
