+++ adlfs/spec.py
@@ -19,17 +19,14 @@ from azure.core.exceptions import (
     ResourceExistsError,
     ResourceNotFoundError,
 )
-from azure.datalake.store import AzureDLFileSystem, lib
-from azure.datalake.store.core import AzureDLFile, AzureDLPath
 from azure.storage.blob import BlobSasPermissions, generate_blob_sas
 from azure.storage.blob._models import BlobBlock, BlobProperties, BlobType
 from azure.storage.blob._shared.base_client import create_configuration
 from azure.storage.blob.aio import BlobServiceClient as AIOBlobServiceClient
 from azure.storage.blob.aio._list_blobs_helper import BlobPrefix
-from fsspec import AbstractFileSystem
 from fsspec.asyn import AsyncFileSystem, get_loop, get_running_loop, sync, sync_wrapper
 from fsspec.spec import AbstractBufferedFile
-from fsspec.utils import infer_storage_options, tokenize
+from fsspec.utils import infer_storage_options
 
 from .utils import (
     close_container_client,
@@ -64,230 +61,6 @@ VERSIONED_BLOB_PROPERTIES = [
 _ROOT_PATH = "/"
 
 _SOCKET_TIMEOUT_DEFAULT = object()
-
-
-class AzureDatalakeFileSystem(AbstractFileSystem):
-    """
-    Access Azure Datalake Gen1 as if it were a file system.
-
-    This exposes a filesystem-like API on top of Azure Datalake Storage
-
-    Parameters
-    -----------
-    tenant_id:  string
-        Azure tenant, also known as the subscription id
-    client_id: string
-        The username or serivceprincipal id
-    client_secret: string
-        The access key
-    store_name: string (optional)
-        The name of the datalake account being accessed.  Should be inferred from the urlpath
-        if using with Dask read_xxx and to_xxx methods.
-
-    Examples
-    --------
-
-    >>> adl = AzureDatalakeFileSystem(tenant_id="xxxx", client_id="xxxx",
-    ...                               client_secret="xxxx")
-    >>> adl.ls('')
-
-    Sharded Parquet & CSV files can be read as
-
-    >>> storage_options = dict(tennant_id=TENNANT_ID, client_id=CLIENT_ID,
-    ...                        client_secret=CLIENT_SECRET)  # doctest: +SKIP
-    >>> ddf = dd.read_parquet('adl://store_name/folder/filename.parquet',
-    ...                       storage_options=storage_options)  # doctest: +SKIP
-
-    >>> ddf = dd.read_csv('adl://store_name/folder/*.csv'
-    ...                   storage_options=storage_options)  # doctest: +SKIP
-
-
-    Sharded Parquet and CSV files can be written as
-
-    >>> ddf.to_parquet("adl://store_name/folder/filename.parquet",
-    ...                storage_options=storage_options)  # doctest: +SKIP
-
-    >>> ddf.to_csv('adl://store_name/folder/*.csv'
-    ...            storage_options=storage_options)  # doctest: +SKIP
-    """
-
-    protocol = "adl"
-
-    def __init__(self, tenant_id, client_id, client_secret, store_name):
-        super().__init__()
-        self.tenant_id = tenant_id
-        self.client_id = client_id
-        self.client_secret = client_secret
-        self.store_name = store_name
-        self.do_connect()
-
-    @staticmethod
-    def _get_kwargs_from_urls(paths):
-        """Get the store_name from the urlpath and pass to storage_options"""
-        ops = infer_storage_options(paths)
-        out = {}
-        if ops.get("host", None):
-            out["store_name"] = ops["host"]
-        return out
-
-    @classmethod
-    def _strip_protocol(cls, path):
-        ops = infer_storage_options(path)
-        return ops["path"]
-
-    def do_connect(self):
-        """Establish connection object."""
-        token = lib.auth(
-            tenant_id=self.tenant_id,
-            client_id=self.client_id,
-            client_secret=self.client_secret,
-        )
-        self.azure_fs = AzureDLFileSystem(token=token, store_name=self.store_name)
-
-    def ls(self, path, detail=False, invalidate_cache=True, **kwargs):
-        files = self.azure_fs.ls(
-            path=path, detail=detail, invalidate_cache=invalidate_cache
-        )
-
-        for file in (file for file in files if type(file) is dict):
-            if "type" in file:
-                file["type"] = file["type"].lower()
-            if "length" in file:
-                file["size"] = file["length"]
-        return files
-
-    def info(self, path, invalidate_cache=True, expected_error_code=404, **kwargs):
-        info = self.azure_fs.info(
-            path=path,
-            invalidate_cache=invalidate_cache,
-            expected_error_code=expected_error_code,
-        )
-        info["size"] = info["length"]
-        """Azure FS uses upper case type values but fsspec is expecting lower case"""
-        info["type"] = info["type"].lower()
-        return info
-
-    def _trim_filename(self, fn, **kwargs):
-        """Determine what kind of filestore this is and return the path"""
-        so = infer_storage_options(fn)
-        fileparts = so["path"]
-        return fileparts
-
-    def glob(self, path, details=False, invalidate_cache=True, **kwargs):
-        """For a template path, return matching files"""
-        adlpaths = self._trim_filename(path)
-        filepaths = self.azure_fs.glob(
-            adlpaths, details=details, invalidate_cache=invalidate_cache
-        )
-        return filepaths
-
-    def isdir(self, path, **kwargs):
-        """Is this entry directory-like?"""
-        try:
-            return self.info(path)["type"].lower() == "directory"
-        except FileNotFoundError:
-            return False
-
-    def isfile(self, path, **kwargs):
-        """Is this entry file-like?"""
-        try:
-            return self.azure_fs.info(path)["type"].lower() == "file"
-        except Exception:
-            return False
-
-    def _open(
-        self,
-        path,
-        mode="rb",
-        block_size=None,
-        autocommit=True,
-        cache_options: dict = {},
-        **kwargs,
-    ):
-        return AzureDatalakeFile(self, path, mode=mode)
-
-    def read_block(self, fn, offset, length, delimiter=None, **kwargs):
-        return self.azure_fs.read_block(fn, offset, length, delimiter)
-
-    def ukey(self, path):
-        return tokenize(self.info(path)["modificationTime"])
-
-    def size(self, path):
-        return self.info(path)["length"]
-
-    def rmdir(self, path):
-        """Remove a directory, if empty"""
-        self.azure_fs.rmdir(path)
-
-    def rm_file(self, path):
-        """Delete a file"""
-        self.azure_fs.rm(path)
-
-    def __getstate__(self):
-        dic = self.__dict__.copy()
-        logger.debug("Serialize with state: %s", dic)
-        return dic
-
-    def __setstate__(self, state):
-        logger.debug("De-serialize with state: %s", state)
-        self.__dict__.update(state)
-        self.do_connect()
-
-
-class AzureDatalakeFile(AzureDLFile):
-    # TODO: refoctor this. I suspect we actually want to compose an
-    # AbstractBufferedFile with an AzureDLFile.
-
-    def __init__(
-        self,
-        fs,
-        path,
-        mode="rb",
-        autocommit=True,
-        block_size=2**25,
-        cache_type="bytes",
-        cache_options=None,
-        *,
-        delimiter=None,
-        **kwargs,
-    ):
-        super().__init__(
-            azure=fs.azure_fs,
-            path=AzureDLPath(path),
-            mode=mode,
-            blocksize=block_size,
-            delimiter=delimiter,
-        )
-        self.fs = fs
-        self.path = AzureDLPath(path)
-        self.mode = mode
-
-    def seek(self, loc: int, whence: int = 0, **kwargs):
-        """Set current file location
-
-        Parameters
-        ----------
-        loc: int
-            byte location
-
-        whence: {0, 1, 2}
-            from start of file, current location or end of file, resp.
-        """
-        loc = int(loc)
-        if not self.mode == "rb":
-            raise ValueError("Seek only available in read mode")
-        if whence == 0:
-            nloc = loc
-        elif whence == 1:
-            nloc = self.loc + loc
-        elif whence == 2:
-            nloc = self.size + loc
-        else:
-            raise ValueError("invalid whence (%s, should be 0, 1 or 2)" % whence)
-        if nloc < 0:
-            raise ValueError("Seek before start of file")
-        self.loc = nloc
-        return self.loc
 
 
 # https://github.com/Azure/azure-sdk-for-python/issues/11419#issuecomment-628143480
